Suppress and fix a few warnings

+++ test/interactive.hpp
@@ -26,7 +26,7 @@ void interactive(int argc, char** argv) {
 		int n = atoi(argv[1]);
 		keys.reserve(n);
         util::XorShift64 prng(0x5603141978c51071);
-		for (uint64_t i = 0; i < n; i++) keys.push_back(bez::function::hash128_t(prng(), prng()));
+		for (int i = 0; i < n; i++) keys.push_back(bez::function::hash128_t(prng(), prng()));
 	} else {
 		for (std::string key; getline(std::cin, key) && key != "";) {
 			strings.push_back(key);
